+++ src/delay.js
@@ -1,0 +1,7 @@
+export function delay(count) {
+    return new Promise((resolve) => {
+        setTimeout(() => {
+            resolve("Loading...")
+        },count)
+    })
+}
